task_10/requirements.txt
fastapi
sqlmodel
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-dotenv
